Move os/path.pyi to posixpath.pyi. Make os.pyi use the latter.
