+++ stdlib/2.7/os/__init__.pyi
@@ -4,7 +4,7 @@ from typing import (
     List, Tuple, Union, Sequence, Mapping, IO, Any, Optional, AnyStr, Iterator,
     MutableMapping, NamedTuple
 )
-from os import path
+import posixpath as path
 
 error = OSError
 name = ... # type: str
